restructuring and adding SDL2_Image
