+++ .gitmodules
@@ -1,7 +1,10 @@
 [submodule "E2DE/Vendor/ECS/EnTT"]
 	path = E2DE/Vendor/ECS/EnTT
 	url = https://github.com/skypjack/entt.git
-[submodule "E2DE/Vendor/SDL/SDL2"]
-	path = E2DE/Vendor/SDL/SDL2
+[submodule "E2DE/Vendor/SDL/SDL_CORE/SDL2"]
+	path = E2DE/Vendor/SDL/SDL_CORE/SDL2
 	url = https://github.com/libsdl-org/SDL.git
 	branch = SDL2
+[submodule "E2DE/Vendor/SDL/SDL_IMAGE/SDL_IMG"]
+	path = E2DE/Vendor/SDL/SDL_IMAGE/SDL_IMG
+	url = https://github.com/libsdl-org/SDL_image.git
