+++ Demo/build-d.sh
@@ -4,4 +4,6 @@ premake5 gmake2
 
 make -C build/
 
+clear
+
 ./bin/Debug/Demo
