+++ E2DE/src/Engine/Texture.h
@@ -1,0 +1,8 @@
+#ifndef E2E_TEXTURE_H_
+#define E2E_TEXTURE_H_
+
+namespace e2e{
+    
+}
+
+#endif
