+++ E2DE/src/Math/Vector.cpp
